% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabelaBivariada.R
\name{TabelaBivariada}
\alias{TabelaBivariada}
\title{Cria tabela de frequências para duas variáveis}
\usage{
TabelaBivariada(w, z, arg = 1)
}
\description{
Cria tabela de frequência absoluta para arg=1,frequência relativa para arg=2, proporção por linha para arg=3 e proporção por coluna para arg=4
}
\details{
@param y Vetor fator ou vetor numérico

@return Matriz

@examples
 TabelaBivariada(mtcars$am, mtcars$carb, 4)
 TabelaBivariada(mtcars$hp, mtcars$am, 2)
 TabelaBivariada(mtcars$hp, mtcars$mpg)

@export
}
