% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabelaFrequenciaSimples.R
\name{TabelaFrequenciaSimples}
\alias{TabelaFrequenciaSimples}
\title{Cria tabela de frequências para uma variável de tipo qualitativa ou quantitativa}
\usage{
TabelaFrequenciaSimples(y)
}
\description{
Cria tabela de frequências de acordo com o tipo de variável testada
}
\details{
@param y Vetor fator não-ordenado, fator ordenado ou numérico

@return Matriz

@examples
 aux <- mtcars$wt
 TabelaFrequenciasimples(aux)
 var <- mtcars$cyl
 TabelaFrequenciasimples(var)

@export
}
