% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabelaNominal.R
\name{TabelaNominal}
\alias{TabelaNominal}
\title{Cria tabela de frequências para variável de tipo nominal}
\usage{
TabelaNominal(y)
}
\description{
Cria tabela de frequência absoluta e frequência relativa com totais por coluna para uma variável de tipo nominal
}
\details{
@param y Vetor fator não-ordenado

@return Matriz

@examples
 aux <- mtcars$vs
 TabelaNominal(aux)

@export
}
