% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabelaux.R
\name{Tabelaux}
\alias{Tabelaux}
\title{Cria tabela de frequências}
\usage{
Tabelaux(y)
}
\description{
Cria tabela defrequência absoluta e frequência relativa
}
\details{
@param y Vetor fator

@return Matriz

@examples
 aux <- mtcars$vs
 Tabelaux(aux)

@export
}
